clean redux
