--- src/redux/index.js
@@ -1,2 +1,0 @@
-export { logIn, logOut } from "./users/usersActions";
-export { getSessions, createSessions } from "./sessions/sessionsActions";
