--- src/redux/sessions/usersActions.js
@@ -1,14 +1,0 @@
-import { LOG_IN } from "./usersType";
-import { LOG_OUT } from "./usersType";
-
-export const logIn = () => {
-	return {
-		type: LOG_IN,
-	};
-};
-
-export const logOut = () => {
-	return {
-		type: LOG_OUT,
-	};
-};
