--- src/redux/users/usersType.js
@@ -1,2 +1,0 @@
-export const LOG_IN = "LOG_IN";
-export const LOG_OUT = "LOG_OUT";
